app/requirements.txt
flask==2.3.3
transformers==4.38.2
tensorflow-cpu==2.15.0
emoji==2.10.0
pandas==2.2.2
gunicorn==21.2.0